feat: Enable interactive selection of map segments
Introduces state and logic to allow users to interactively select segments on the map, in addition to points.

To provide a clearer user experience, the selection of points and segments is now mutually exclusive. Selecting a segment will deselect any active point, and vice-versa. This is managed by passing new state and a selection handler to the map component.